Add tyr-cli module
