+++ pom.xml
@@ -73,6 +73,7 @@
 
     <modules>
         <module>tyr-api</module>
+        <module>tyr-cli</module>
         <module>tyr-core</module>
         <module>tyr-webhook</module>
         <module>webhook-load-balancer</module>
