+++ tyr-cli/src/main/resources/application.properties
@@ -1,0 +1,1 @@
+quarkus.native.additional-build-args = -H:ReflectionConfigurationFiles=reflection-config.json
